+++ packages/web/src/app/[lng]/_components/index.ts
@@ -1,3 +1,3 @@
-export { default as Header } from './Header';
+export { default as Header } from '../../../components/Header';
 export { default as ThemeProvider } from './ThemeProvider';
 export { default as QueryProvider } from './QueryProvider';
